requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
Django==3.2.3
django-bootstrap4==3.0.1
django-crispy-forms==1.11.2
django-fontawesome-5==1.0.18
install==1.3.4
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
